pyproject.toml
[project]
name = "acessorIA"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "google-genai>=1.68.0",
    "langchain>=1.2.14",
    "langchain-google-genai>=4.2.1",
    "pymongo>=4.16.0",
    "langchain-groq>=1.1.1",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "langchain-text-splitters>=1.1.2",
    "faiss-cpu>=1.13.2",
    "langchain-community>=0.4.1",
    "pypdf>=6.10.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pre-commit>=4.5.0"
]

[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra -q"
testpaths = ["tests"]
pythonpath = ". src"

[tool.black]
target-version = ['py312']
line-length = 88
